chore(core): move collateralization ratio display
